fix(review): apply post-review fixes across domain, UI, and tests
- Rename MoveRotation.double → MoveRotation.half (shadowed Dart's double type);
  update spec section 2.2 and all call sites
- Add centre-uniqueness check to CubeValidator (D4: impossible states rejected)
- Disable Solve/Teach/Scramble/Reset buttons while _isSolving in EditorScreen
- Remove duplicate _CubeLayout from cube_screen.dart; use shared CubeLayout widget
- Fix cube_validator_test.dart: add missing test() header; update error-count
  expectation from 6 → 11 to account for new centre-uniqueness errors

+++ lib/core/cube.dart
@@ -93,7 +93,7 @@ class Cube {
     final reps = switch (move.rotation) {
       MoveRotation.cw => 1,
       MoveRotation.ccw => 3,
-      MoveRotation.double => 2,
+      MoveRotation.half => 2,
     } *
         times;
     var cube = this;
